backend/requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
python-dotenv
reportlab
python-pptx
openpyxl
matplotlib
pytube
Pillow
moviepy
duckduckgo-search
youtube-transcript-api
openai
jinja2
sumy
PyPDF2
python-docx
reportlab